+Added read_variable and read_attribute to MOM_io
  Added the new set of routines read_variable to do a simple read of an entire
array from a netCDF file with the root PE, and then broadcast this information
to the other PEs.  Also added read_attribute to read a global or variable
attribute.  The option to read dimension names was added to get_var_sizes, and
the option for get_varid to indicate whether a variable has been found. As a
part of this, separate variants of broadcast were added for 32-bit and 64-bit
integers, as were new variants of MOM_read_data for 1-d and 0-d integer fields.
All answers are bitwise identical, but there are new options and optional
arguments for overloaded interface, and there are two new overloaded interfaces
for reading.