+read_attribute uses allocatable character string
  Modified read_attribute_str to work with allocatable character string
arguments to properly handle string attributes of unknown (and possibly
enormous) size, following advice from Keith Lindsay.  Because this is a change
in an argument to a (recently added) publicly visible interface, it required
changes to the two places where it was called.  All answers are bitwise
identical, but an interface has changed.