+Add ncid_in optional args to MOM_io read routines
  Added the option to provide a handle to an open netCDF file via the new
ncid_in optional arguments to various routines that read from netCDF files.
Along with the new interfaces open_file_to_read and close_file_to read in MOM_io
these allow for files to be opened and closed once while reading a number of
fields or attributes from the same file.  If these arguments are not provided,
the files are opened and closed with each call as before.  All answers are
bitwise identical, although there are new optional arguments and new public
interfaces.