NCAR sponge merge (#1308)
This PR contains updates provided by NCAR that act on velocities in sponges to provide damping accelerations toward specified velocities in MOM_ALE_sponge.  Associated with this change are new runtime parameters to control the use of sponges on tracers and new diagnostics of the tendencies due to the ALE sponges, but these changes only appear in the MOM_parameter_doc or available_diags files if ALE sponges are in active use or if the sponges are initialized from a file, and no changes to these files were found with the standard pipeline testing, perhaps reflecting an important limitation in this testing.

As a part of this work, a bug was identified in the case of incoming sponge data residing on the model horizontal grid
where the returned mask was not initialized properly. The 2018 answers flags are being used to retain the bug and should be set to False for new experiments.

The new velocity sponges are activated by setting SPONGE_UV=True and SPONGE_CONFIG="file".
The sponge accelerations use the tracer damping timescale by default but this can be set independently.

Significant pre-squash commit descriptions include:
* sponge layer changes

* add uv-specific iresttime for sponges

* fix idamp_u init

* fix mask_z init

* improve tracer sponge apply

* fix velocity sponge apply

* Enable sponge tendency diagnostics
  - sp_tendency_temp and sp_tendency_salt are new diagnostic
    variables which evaluate the respective tracer tendendies (tr_units/sec).
  - sp_tendency_u and sp_tendency_v diagnose the accelerations
    (m/sec) applied from calls to the sponge routine.
  - No attempt has been made to CMOR-ize these diagnostics.
  - More work is needed to generalize this code for additional tracers.

* merging updates from NCAR for uv sponges

* Changes for uv sponges and sponge diagnostics

  - this PR includes updates from NCAR for uv momentum
    sponge implementation.
  - Diagnostics are included for tracer tendencies and
    accelerations due to sponge terms.
  - The uv sponge feature is currently not being tested. This will be
    addressed in a future PR which will add sponge accelerations
    to tc4.

* initialize h_col prior to calling remapping

* Toggle bug in horiz_interp_and_extrap_tracer mask for ongrid data

  - A previous bug in the 3-dimensional mask returned from this
    routine in the case where the data to be interpolated reside
    on the model's horizontal grid is retained using the 2018_answers flag.
  - If 2018_answers is set to True, the mask is not properly initialized
    and this leads to incorrect vertical reconstructions in the ALE
    sponge code.
  - 2018 flags (DEFAULT_2018_ANSWERS or HOR_REGRID_2018_ANSWERS)
    should be set to False for any cases using sponge restoring for
    tracers or momentum.

* update parameter documentation

* Adds uv sponge initialization for non time-varying data

Co-authored-by: alperaltuntas <alperaltuntas@gmail.com>
Co-authored-by: Robert Hallberg <Robert.Hallberg@noaa.gov>