+Added open_ASCII_file
  Added the new subroutine open_ASCII_file to open text files that can be read
or written to using simple Fortran read or write statements.  This was
previously handled via a common open_file call, but FMS2 is changing the
interface for this routine, and creating this new file de-conflicts most the
MOM6 calls to open_file from these altered interfaces.  The existing routine
open_file is unaltered and works exactly as before.  A number of existing calls
to open_file were also modified to use this new interface, in some cases where
only the root_PE was writing, this includes altering the code so that only the
root_PE actually opens the file.  All answers and output are bitwise identical.