+Revised the depth list type for standard I/O
  Revised the depth list type used in MOM_sum_output to accommodate I/O using
ordinary routines, and to make the code creating, reading and writing it
potentially separable from the MOM_sum_output module by avoiding the use of the
MOM_sum_output control structure in most of the routines.  Also changed the
index 'l' to 'li' in some places to avoid having it misread as '1'.  All answers
are bitwise identical but there are changes in some subroutine arguments.