Remove AC_CHECK_FILE for legacy MOM_memory.h
This patch removes the AC_CHECK_FILE autoconf macro which assigns a
default path to the MOM_memory.h file path.

This path was defaulting back to the old directory, and was added to
support the regression verification testing, but this is a regression
problem with should be handled by the regression, not the autoconf
configuration.

This will produce a regression test fail, but it not a cause for concern.