Correct module use statements in infra/FMS1
  Corrected three module use statements in infra/FMS1 to eliminate any
dependencies of code in config_src/infra/FMS1 on code in src/framework.  These
use statements eventually point to the same place as before, but with less
indirection.  This change should facilitate later steps to compile everything
in and under the config_src/infra directories as libraries.  All answers are
bitwise identical.