Bugfix: FGNV streamfunction vertical bounds
This patch fixes an issue with the vertical array bounds of the Ferrari
et al. streamfunction.  The array is bounded across interfaces, from 1
to nz+1, but only the interior values need to be determined due to the
arbitrary boundary value (set here to zero).

In the current source, the streamfunction is rescaled before calling
streamfn_solver, but need not be applied to the boundary values.  This
is unlikely to cause errors in production, since the values are later
reset to zero, but the rescaling can raise errors in more aggressive
debugging builds, such as when the arrays are initialized with NaN
values.