+Simplified read_data_fms2 and write_field_fms2
  Eliminated scale_data from MOM_read_data_fms2, as it was duplicative of
rescale_comp_data from MOM_domain_infra.F90.  Also eliminated the optional scale
argument from the fms2 write_field routines, because this scaling is already
handled in the framework/MOM_io.F90 routines, and because the way the scaling
was implemented with a pointer to defeat the intent(in) of the data array to
the write routines was a deviously terrible idea that would cause endless
problems.  Because the mpp_domain in the MOM_domain_types have been modified to
always have an io_domain set, the test and calls setting an io_domain in all of
the fms2 read_data and fms2 write_data routines.  In addition, there is no
longer a need for a routine to read specifically from the supergrid, so the fms2
MOM_read_data_2d_supergrid routine was eliminated.  The code compiles with these
changes, but none of this fms2 I/O code is exercised yet in MOM6, so of course
all answers are bitwise identical.