Clean up in MOM_restart.F90
  Cleaned up MOM_restart.F90 in the wake of the recent merge of FMS2 I/O related
changes into MOM6.  The changes include calling the recently added functions
get_variable_byte_size and get_num_restart_files, and revising the names of
several variables in open_restart_units and the order in which the count of
files is incremented for greater clarity.  There are also a number of
duplicative comments that were removed while others were reformatted.  All
answers and output are bitwise identical.