Only write checksums from root PE for FMS_cap
  Modified the routines that write out the checksums of various types related to
the surface ocean state so that only the root PE writes the checksums.  The
modified routines are ocn_ice_bnd_type_chksum and ocean_public_type_chksum in
config_src/infra/FMS_cap.  The analogous routines with the nuopc_cap and mct_cap
have not been modified; although doing so would be a very good idea, it should
be done by someone who is actively testing them.  In some large-PE tests cases
this reduces the volume of output to stdout by over 90% without any loss of
information.  All solutions are bitwise identical, but there may be minor
white-space reformatting of the output in some cases.