+Add edge_axis argument to write_metadata_axis
  Added a new optional logical argument to write_metadata_axis to indicate when
an axis is staggered at the edges of the tracer grid, and changed calls to
get_file_info to stop requesting the number of global attributes.  Also
eliminated some unused optional arguments to the FMS1 version of write_metadata
field.  All answers are bitwise identical, but there are minor changes to
some I/O related interfaces.