Add missing ".nc" to FMS2 output filenames
  Add a missing ".nc" suffix to the output filename with FMS2_io, while also
issuing a warning, following the practice of FMS1.  Also reordered the calls to
add the longname and axis attributes to FMS2 files, to follow the order used in
MOM6 calls to FMS1.  All answers are bitwise identical, but there are some
changes to output filenames and orders of attributes in files (to revert to
traditional behavior).