Explicit domain decomposition of horizontal axes
FMS2 restart routines expect axes to be domain-decomposed.  However, the
domain_write_1d function does not apply this decomposition and instead
routes this operation to compressed_write_1d.

In order to accommodate this, we explicitly slice the 1d arrays of any
axes into its domain-decomposed segment before passing to write_data.
We have also introduced a control flag to MOM's FMS2 axistype to direct
MOM_write_axis when this needs to be applied.

We currently apply the domain decomposition flag to all horizontal axes
regardless of circumstances.  For now this is probably sufficient, but
may need further testing (e.g. cube sphere).