+More consistent treatment of input_filename = 'F'
  Restructured the code slightly so that the output files that are generated
when input_filename = 'F' is exactly the same as if it is 'n' if there are no
restart files in the restart_input_dir, or as if it is 'r' if the restart files
are there.  Previously, the solutions with 'F' worked this way, but no
ocean_geometry.nc or Vertical_grid.nc files were written when WRITE_GEOM=1,
regardless of the presence or absence of the restart files, and the
MOM_parameter_doc.all files differed slightly between the 'n' and 'F' or 'r'
cases.  As a part of these changes, the determination of whether this is a new
run is moved earlier in the algorithm, and now sits outside of
MOM_initialize_state.  All solutions are bitwise identical, but there are
changes in the position of the PARALLEL_RESTARTFILES and REFERENCE_HEIGHT
entries in most MOM_parameter_doc.all files.