Fix for use of optional argument without checking for presence.
The optional argument answers_2018 is used in calls to fill_miss_2d in
two places but it may not be present to begin with.
As the default in fill_miss_2d is to have a true value if that argument
in not present, I have changed the pass through argument to ans_2018.

Also needed to add ans_2018 and related logic to the horiz_interp_and_extrap_tracer_record
subroutine.