+Add optional global_file argument to MOM_read_data
  Added the optional argument global_file to all of the overloaded MOM_read_data
routines, to indicate whether the file being read is global or is domain-
decomposed using the IO-layout.  With FMS2, this change does nothing, but with
FMS1 it is needed to correctly read from decomposed files that might include 4-d
arrays.  With this change there are versions of SIS2 that reproduce across
restarts without using the FMS1 restart routines.  All answers are bitwise
identical, but there are new optional arguments.