read_attribute_str returns cleanly when found=F
Updates read_attribute_str() so that if the optional argument found is provided
and the attribute is not defined, then the function returns with found=.false.
instead of calling MOM_error(FATAL).

I did a little bit of testing, and it looks like read_attribute_real and
read_attribute_int (not sure if the default is int32 or int64) already behave
correctly when found is present so they do not need a similar update