(*)Correct inconsistent units in 17 diagnostics
  Corrects inconsistent units or unit conversion factors for 17 diagnostics
plus all of the diagnostics for extra tracer fluxes where the scaling factor is
not explicitly supplied (temperature and salinity are not impacted).  In most
cases the corrections are in obscure diagnostics that are probably not used much
or only apply to non-Boussinesq cases.  In addition, dimensional rescaling in
time was applied to the diapycnal velocity diagnostic in layered mode.  This
commit addresses the specific problems noted in MOM6 issue #1384, but it does
not uniformly change to the suggested code construct for register_diag_field
arguments that would make these problems easier to detect in the future.  All
solutions are bitwise identical, but there are changes to the metadata of some
diagnostics and there are changes to the available_diags files.