.testing: Create python venv at compile time
This patch shifts the creation of the python virtual environment to a
compile-time operation.

The .testing suite creates a Python virtual environment (venv) if the
current environment does not include numpy and the Python netCDF4
modules.

These are detected and set up at runtime rather than compile time, which
can cause issues when run on compile nodes which do not have internet
access since the packages need to be downloaded via Pip.

This patch moves the detection and creation to compile time.  Since
internet access is expected for FMS and mkmf setup, it is consistent to
assume internet access for pip installations.