Use MOM_io interfaces in write_depth_list
  Revised write_depth_list to use the interfaces from MOM_io instead of direct
calls to NF90_ routines.  The depth list files that are generated this way are
interchangeable with those from previous versions, but they are larger because
FMS insists on including an unnecessary axis variable with a simple list of
integers for the list axis.  This commit partially addresses MOM6 issue #1312.
All answers are bitwise identical.