+Add dimensional rescaling of wave variables
  Added dimensional rescaling for all of the variables in the MOM_wave,
including rescaling of the Stokes velocities, which are used in two modules
outside of the MOM_wave_interface module.  Also added comments describing
numerous variables inside of the wave module, and removed the public declaration
for most of the elements of the wave_parameters_CS.  All answers are bitwise
identical in tests, but the waves module is not as extensively tested as it
should be.  There are changes in the rescaled units in publicly visible
variables.