infra, FMS_cap: stdout_if_root()
This patch adds `stdout_if_root` to the MOM_io API, and points to the
FMS `stdout()` function in the FMS infra implementations.

This change was required because calls to `coupler_type_write_chksums`
handles both its own checksums across ranks and its own IO to `outunit`.
Typically only the root PE will write the result.

The FMS `stdout()` function would return the designated stdout unit for
the root PE, and the internal `etc_unit` for other PEs, usually set to
`/dev/null`.

When MOM_io switched from using the FMS `stdout()` function to the
`stdout` unit as defined in `iso_fortran_env`, this functionality was
lost and every PE would write the same result to stdout.

Normally this was controlled with `if (root)`-like tests, but this
cannot be used in functions like `coupler_type_write_checksums`, which
require participation of all PEs.

We decided the only resolution here was to introduce a new function,
`stdout_if_root` which replicates the original behavior of `stdout()`.

Additional comments:

* The `if (root)` checks were retained, since it is presumably still a
  good idea to avoid the `write()` calls when possible, even to
  `/dev/null`.

* This patch only updates the FMS coulper driver, but I would recommend
  that the other driver maintainers review their own calls to this
  function.