+Move wave module variables into control structure
  Moved 15 module variables into wave_parameters_CS, to follow the coding style
and conventions in the rest of the code.  Also added comments describing 8
enumeration parameters in this same module, and added comments suggesting ways
that the accuracy of the calculations could be improved.  Also incorporated a
factor of 2*PI into the definition of Freq_Cen in this module.  In addition, the
initialization procedure for this module was altered in some more commonly used
cases, so that use MOM_wave_interface_init to set wave parameters, even with a
statistical wave model, and eliminated MOM_wave_interface_init_lite.  I believe
that all solutions should be bitwise identical, but this module is not yet well
tested in the MOM6 regression tests.