+Avoid using field_exists in MOM_wave_interface
  Modified get_var_sizes so it does return ndims = -1 when querying for a
variable that does not exist in a file (which was the intended behavior) and
made use of this new and proper functionality to avoid using the function
field_exists() in Surface_Bands_by_data_override(), because the fms_io
implementation of field_exists incorrectly returns false when the variable being
sought is also the name of a coordinate.  Three minor bugs in wave diagnostics
that Brandon Reichl had identified were also fixed, as was a test controlling an
allocate statement.  All answers are bitwise identical in cases that worked
before, and the waves code is now working correctly with all IO calls going via
MOM_io, rather than calling netCDF routines directly.