Remove duplicated solvers between regrid modules
In the removal of regrid_edge_slopes, a couple of the solvers became
duplicated between regrid_edge_values and regrid_solvers. Here,
they're simply made consistent by removing the solvers from
regrid_edge_values, and bringing over the zero-pivot check for the
linear solver that was in one variation but not the other.