Autoconf: FMS 2019.01.02 test; framework flag
The autoconf build was updated to verify that calls to diag_axis_init
support the domain_position argument.  This was introduced in FMS
2019.01.02, so this acts as an implicit minimum FMS version test.

This test is done indirectly by confirming that the valid
domain_position values (NORTH, EAST, CENTER) are in diag_axis_mod.

A `--with-framework=` flag was also added to select either the FMS1 or
FMS2 backend.