Diag mediator: linked list for auxiliary axes
During diagnostic registration, the associated axes group (axes_grp) is
tested against the standard set of grid axes groups.  If it matches one
of them, then the axes argument points to the corresponding axes_grp.

If absent, the core axes_grp fields are re-allocated as an internal copy
and passed to the lower level diagnostic registration functions.

However, the current implementation does not track these auxiliary axes
and were a source of memory leak.

This patch introduces an "auxiliary axis" linked list which stores these
non-standard axes_grp copies.  If a diagnostic's axes_grp is not in the
standard set, then it is tested against the auxiliary axis group list.
If not in the list, then it is copied and appended appended.

This approach copies the input axes_grp content to a new allocation, so
it does not address the construction and management of the original
axes_grp.

Also, this current implementation does not copy the contents stored in
axes_grp pointers.  Instead the pointers are copied.  (I guess this is
kind of dumb and needs to be fixed...)