Update to allow data_override for prcme_adj
Removed call to data_override_init. It is called in coupler_init after
call to ocean_driver_init.
The domain that is passed to the MOM version of data_override_init is
ignored after the second call in coupler_init.

Updated the size of the array being used to hold the data being read in.
The domain that was passed in had non-zero haloes while data_override
"sees" a data domain that is the same size as the compute domain.
The G%domain argument passed when reading the data for hflx_adj,
sflx_adj and prcme_adj is not actually used when we get to the
infrastructure version of data_override. It keys off the OCN entry in
the data_table to find the domain (which is what was initialized from
the call from coupler_init).