Allocate Waves%WaveNum_Cen before reading into it
  Moved the recently added call to read SURFBAND_WAVENUMBERS into an array when
SURFBAND_SOURCE=COUPLER down by a line to follow the allocate call for that
array.  This change should avert a memory access problem that would otherwise
arise when exercising this newly added code.  All answers are bitwise identical
in any case that ran with the previous version, and they should reproduce
answers from before this PR as a whole, although obviously this code is not
as well tested as would be ideal, based on the fact that this bug was in the
last commit, which also passed testing.