+Add query_ocean_state
  Added the new routine query_ocean_state to mom_ocean_model_nuopc to allow
for the wave properties to be obtained by the mom_cap without having to rely on
the elements of an otherwise opaque type being public.  I believe that all
elements of the ocean_state_type could now be declared as private, and that the
version of MOM6 with the NUOPC coupler should now work as intended.