Fix to codecov file path search
The codecov.io upload script appears to rely on the filepath of source
codes at compile time, and was unable to find the output when called
from .testing or the work directories.  In this patch, we now move to
`.testing/build/symmetric` (the gcov-enabled build) before running the
uploader script.

This is possibly required after the recent security alert connected to
the codecov.io uploader script.  Sadly, our coverage report failures
went undetected for several months.

The `-Z` flag has also been added to the script, which returns a nonzero
error code if it fails.  It's not yet clear to me if a missing file
registers as a fail, but it may help to detect future problems.