Testing: Rotation support for MOM_read_data
This patch adds an interface and several implementations of
`MOM_read_data` and `MOM_read_vector` functions which support rotational
testing.

The MOM_domain_type now carries a scalar, `turns`, indicating the number
of quarter-turns from the input grid to the model grid.  If this number
is nonzero, then a field is read into an array based on the input grid
and rotates the field to a new array based on the model grid.  This
final result is returned by the function.

If the domain's `turns` is zero, then it is assumed to be a call from a
non-rotated domain and no rotation is applied.  Functions outside of MOM
(such as calls within drivers) do not apply this rotation.

For the "domain-less" reads of 2d arrays, an explicit `turns` argument
is supported.  This only appears to be necessary in one part of grid
initialization.

This is now the third place where `turns` is tracked: the first is HI
(horizontal index tracking) of the MOM grid, the second in `restart_CS`,
and now in `MOM_Domain`.  However, I believe this is a reasonable (if
not necessary) place to track the domains while also avoiding a need for
users to explicitly rotate fields every time `read_data` is called.