(*)Fix discretization issues with USE_GME=True
  Corrected the inconsistently staggered expressions for the total depths used
to scale away the GM+E backscatter scheme that is enabled with USE_GME.  As a
part of these changes, the nominal bathymetric depths have been replaced by the
sum of the layer thicknesses, and parentheses were added to some of the GME
expressions so that they will be rotationally invariant.  Although there are no
changes to the answers in the MOM6-examples test suite, answers will change in
any cases where USE_GME is true, and GME_H0 is larger than the minimum ocean
depth.  This commit will address MOM6 issue #1466, which can now be closed.